Fixed getting sample-rates for input/output only device for PortAudio Signed-off-by:doe300 <stadeldani@web.de>

+++ include/processors/AudioProcessor.h
@@ -28,7 +28,7 @@ namespace ohmcomm
         /*!
          * The current time in the stream (i.e. elapsed microseconds)
          */
-        unsigned int streamTime;
+        unsigned long streamTime;
 
         /*!
          * The maximum number of bytes to be stored in the input/output-buffer
